# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ilyas Store eCommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin sees all, user sees their own)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}/status:
    put:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusBody"
      responses:
        "200":
          description: Order updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # USERS
...
  /coupons:
    get:
      operationId: listCoupons
      tags: [coupons]
      summary: List all coupons (admin)
      responses:
        "200":
          description: Coupons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Coupon"
    post:
      operationId: createCoupon
      tags: [coupons]
      summary: Create coupon (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "201":
          description: Coupon created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
  /coupons/{id}:
    put:
      operationId: updateCoupon
      tags: [coupons]
      summary: Update coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCouponBody"
      responses:
        "200":
          description: Coupon updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Coupon"
    delete:
      operationId: deleteCoupon
      tags: [coupons]
      summary: Delete coupon (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /coupons/validate:
    post:
      operationId: validateCoupon
      tags: [coupons]
      summary: Validate a coupon code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ValidateCouponBody"
      responses:
        "200":
          description: Coupon validity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CouponValidationResponse"

  # SETTINGS
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]

    # AUTH
...
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        variantId:
          type: integer
          nullable: true
        productNameSnapshot:
          type: string
        variantValueSnapshot:
          type: string
          nullable: true
        priceSnapshot:
          type: number
        quantity:
          type: integer
      required: [productId, productNameSnapshot, priceSnapshot, quantity]
    OrderAddress:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        street:
          type: string
        city:
          type: string
        country:
          type: string
      required: [name, phone, street, city]
    Order:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        totalPrice:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        address:
          $ref: "#/components/schemas/OrderAddress"
        couponCode:
          type: string
        discount:
          type: number
        createdAt:
          type: string
          format: date-time
      required: [id, userId, items, totalPrice, currency, status, address, createdAt]
    CreateOrderBody:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              variantId:
                type: integer
                nullable: true
              quantity:
                type: integer
            required: [productId, quantity]
        address:
          $ref: "#/components/schemas/OrderAddress"
        currency:
          type: string
          default: PKR
        couponCode:
          type: string
      required: [items, address]
    UpdateOrderStatusBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
      required: [status]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [orders, total, page, totalPages]

    # CART ITEMS
...
    Coupon:
      type: object
      properties:
        id:
          type: integer
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxUses:
          type: integer
        usedCount:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, code, discountType, discountValue, isActive, usedCount, createdAt]
    CreateCouponBody:
      type: object
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed]
        discountValue:
          type: number
        minOrderAmount:
          type: number
        maxUses:
          type: integer
        isActive:
          type: boolean
        expiresAt:
          type: string
          format: date-time
      required: [code, discountType, discountValue]
    ValidateCouponBody:
      type: object
      properties:
        code:
          type: string
        orderAmount:
          type: number
      required: [code, orderAmount]
    CouponValidationResponse:
      type: object
      properties:
        valid:
          type: boolean
        discount:
          type: number
        message:
          type: string
      required: [valid, message]

    # SETTINGS
...